refactored logic with UiTestSettings class
