+++ EpamWebSiteTestingFramework/Epam.TestAutomation.Core/Helper/UiTestSettings.cs
@@ -1,37 +1,29 @@
 using Epam.TestAutomation.Core.Enums;
-using Epam.TestAutomation.Core.Utils;
 using Epam.TestAutomation.TestData;
+using Epam.TestAutomation.Utilities.EnumParser;
 using Epam.TestAutomation.Utilities.JsonParser;
 
 namespace Epam.TestAutomation.Core.Helper;
 
 public static class UiTestSettings
 {
-    private static TestInfo TestInfo { get; set; }
+    private static BrowserType Browser => EnumParser.GetEnumValueByDescription<BrowserType>("DefaultBrowser");
+    private static TestInfo TestInfo => GetTestInfoFromJson();
 
-    public static BrowserType Browser => EnumUtils.GetEnumValueByDescription<BrowserType>("DefaultBrowser");
+    private static TestInfo GetTestInfoFromJson()
+    {
+        var json = File.ReadAllText(
+            "/Users/kate/RiderProjects/Test_Automation_Framework/EpamWebSiteTestingFramework/Epam.TestAutomation.TestData/testdata.json"
+        );
+        return JsonParser.DeserializeJsonToObject<TestInfo>(json);
+    }
 
     public static string ScreenshotPath => TestInfo.ScreenshotPath;
-
+    public static string ScreenshotPath2 => TestInfo.ScreenshotPath2;
     public static string LogsPath => TestInfo.LogsPath;
-
     public static TimeSpan WebDriverTimeOut => TimeSpan.FromSeconds(TestInfo.WebDriverTimeOut);
-
     public static TimeSpan DefaultTimeOut => TimeSpan.FromSeconds(TestInfo.DefaultTimeOut);
+    public static BrowserType GetCurrentBrowser => Browser;
 
-    // TODO refactor method
-    public static string ApplicationUrl()
-    {
-        GetTestInfo();
-        return TestInfo.ApplicationUrl;
-    }
-
-    private static TestInfo GetTestInfo()
-    {
-        var json =
-            File.ReadAllText(
-                "/Users/kate/RiderProjects/Test_Automation_Framework/EpamWebSiteTestingFramework/Epam.TestAutomation.TestData/testdata.json");
-        TestInfo = JsonParser.DeserializeJsonToObject<TestInfo>(json);
-        return TestInfo;
-    }
+    public static string ApplicationUrl() => TestInfo.ApplicationUrl;
 }
