+++ EpamWebSiteTestingFramework/Epam.TestAutomation.TestData/TestInfo.cs
@@ -8,6 +8,7 @@ public class TestInfo
 
     [JsonPropertyName("screenshotPath")] public string ScreenshotPath { get; set; }
 
+    [JsonPropertyName("screenshotPath2")] public string ScreenshotPath2 { get; set; }
     [JsonPropertyName("logsPath")] public string LogsPath { get; set; }
 
     [JsonPropertyName("webDriverTimeOut")] public int WebDriverTimeOut { get; set; }
